+++ tencentcloud/cfg/v20210820/models.go
@@ -1068,6 +1068,14 @@ type TaskListItem struct {
 
 	// 任务更新时间
 	TaskUpdateTime *string `json:"TaskUpdateTime,omitempty" name:"TaskUpdateTime"`
+
+	// 0--未开始，1--进行中，2--已完成
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	TaskPreCheckStatus *int64 `json:"TaskPreCheckStatus,omitempty" name:"TaskPreCheckStatus"`
+
+	// 环境检查是否通过
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	TaskPreCheckSuccess *bool `json:"TaskPreCheckSuccess,omitempty" name:"TaskPreCheckSuccess"`
 }
 
 type TaskMonitor struct {