+++ tencentcloud/essbasic/v20210526/models.go
@@ -1782,10 +1782,10 @@ type DescribeTemplatesRequestParams struct {
 	// 操作者的信息
 	Operator *UserInfo `json:"Operator,omitempty" name:"Operator"`
 
-	// 是否返回所有组件信息
+	// 是否返回所有组件信息。默认false，只返回发起方控件；true，返回所有签署方控件
 	QueryAllComponents *bool `json:"QueryAllComponents,omitempty" name:"QueryAllComponents"`
 
-	// 模糊搜索模板名称
+	// 模糊搜索模板名称，最大长度200
 	TemplateName *string `json:"TemplateName,omitempty" name:"TemplateName"`
 }
 
@@ -1810,10 +1810,10 @@ type DescribeTemplatesRequest struct {
 	// 操作者的信息
 	Operator *UserInfo `json:"Operator,omitempty" name:"Operator"`
 
-	// 是否返回所有组件信息
+	// 是否返回所有组件信息。默认false，只返回发起方控件；true，返回所有签署方控件
 	QueryAllComponents *bool `json:"QueryAllComponents,omitempty" name:"QueryAllComponents"`
 
-	// 模糊搜索模板名称
+	// 模糊搜索模板名称，最大长度200
 	TemplateName *string `json:"TemplateName,omitempty" name:"TemplateName"`
 }