+++ tencentcloud/monitor/v20180724/models.go
@@ -172,6 +172,10 @@ type AlarmNotice struct {
 	// 推送cls渠道
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	CLSNotices []*CLSNotice `json:"CLSNotices,omitempty" name:"CLSNotices"`
+
+	// 通知模版绑定的标签
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 type AlarmPolicy struct {
@@ -309,6 +313,14 @@ type AlarmPolicy struct {
 	// 高级指标数量
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	AdvancedMetricNumber *int64 `json:"AdvancedMetricNumber,omitempty" name:"AdvancedMetricNumber"`
+
+	// 策略是否是全部对象策略
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	IsBindAll *int64 `json:"IsBindAll,omitempty" name:"IsBindAll"`
+
+	// 策略标签
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 type AlarmPolicyCondition struct {
@@ -411,6 +423,14 @@ type AlarmPolicyRule struct {
 	// 集成中心产品ID
 	// 注意：此字段可能返回 null，表示取不到有效值。
 	ProductId *string `json:"ProductId,omitempty" name:"ProductId"`
+
+	// 最大值
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ValueMax *float64 `json:"ValueMax,omitempty" name:"ValueMax"`
+
+	// 最小值
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	ValueMin *float64 `json:"ValueMin,omitempty" name:"ValueMin"`
 }
 
 type AlarmPolicyTriggerTask struct {
@@ -867,6 +887,9 @@ type CreateAlarmNoticeRequestParams struct {
 
 	// 推送CLS日志服务的操作 最多1个
 	CLSNotices []*CLSNotice `json:"CLSNotices,omitempty" name:"CLSNotices"`
+
+	// 模版绑定的标签
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 type CreateAlarmNoticeRequest struct {
@@ -892,6 +915,9 @@ type CreateAlarmNoticeRequest struct {
 
 	// 推送CLS日志服务的操作 最多1个
 	CLSNotices []*CLSNotice `json:"CLSNotices,omitempty" name:"CLSNotices"`
+
+	// 模版绑定的标签
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 func (r *CreateAlarmNoticeRequest) ToJsonString() string {
@@ -913,6 +939,7 @@ func (r *CreateAlarmNoticeRequest) FromJsonString(s string) error {
 	delete(f, "UserNotices")
 	delete(f, "URLNotices")
 	delete(f, "CLSNotices")
+	delete(f, "Tags")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "CreateAlarmNoticeRequest has unknown keys!", "")
 	}
@@ -987,6 +1014,12 @@ type CreateAlarmPolicyRequestParams struct {
 
 	// 聚合维度列表，指定按哪些维度 key 来做 group by
 	GroupBy []*string `json:"GroupBy,omitempty" name:"GroupBy"`
+
+	// 模版绑定的标签
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
+
+	// 日志告警信息
+	LogAlarmReqInfo *LogAlarmReq `json:"LogAlarmReqInfo,omitempty" name:"LogAlarmReqInfo"`
 }
 
 type CreateAlarmPolicyRequest struct {
@@ -1033,6 +1066,12 @@ type CreateAlarmPolicyRequest struct {
 
 	// 聚合维度列表，指定按哪些维度 key 来做 group by
 	GroupBy []*string `json:"GroupBy,omitempty" name:"GroupBy"`
+
+	// 模版绑定的标签
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
+
+	// 日志告警信息
+	LogAlarmReqInfo *LogAlarmReq `json:"LogAlarmReqInfo,omitempty" name:"LogAlarmReqInfo"`
 }
 
 func (r *CreateAlarmPolicyRequest) ToJsonString() string {
@@ -1061,6 +1100,8 @@ func (r *CreateAlarmPolicyRequest) FromJsonString(s string) error {
 	delete(f, "TriggerTasks")
 	delete(f, "Filter")
 	delete(f, "GroupBy")
+	delete(f, "Tags")
+	delete(f, "LogAlarmReqInfo")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "CreateAlarmPolicyRequest has unknown keys!", "")
 	}
@@ -3681,6 +3722,9 @@ type DescribeAlarmNoticesRequestParams struct {
 
 	// 根据通知模板 id 过滤，空数组/不传则不过滤
 	NoticeIds []*string `json:"NoticeIds,omitempty" name:"NoticeIds"`
+
+	// 模版根据标签过滤
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 type DescribeAlarmNoticesRequest struct {
@@ -3715,6 +3759,9 @@ type DescribeAlarmNoticesRequest struct {
 
 	// 根据通知模板 id 过滤，空数组/不传则不过滤
 	NoticeIds []*string `json:"NoticeIds,omitempty" name:"NoticeIds"`
+
+	// 模版根据标签过滤
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 func (r *DescribeAlarmNoticesRequest) ToJsonString() string {
@@ -3739,6 +3786,7 @@ func (r *DescribeAlarmNoticesRequest) FromJsonString(s string) error {
 	delete(f, "UserIds")
 	delete(f, "GroupIds")
 	delete(f, "NoticeIds")
+	delete(f, "Tags")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeAlarmNoticesRequest has unknown keys!", "")
 	}
@@ -3846,6 +3894,15 @@ type DescribeAlarmPoliciesRequestParams struct {
 
 	// 根据一键告警策略筛选 不传展示全部策略 ONECLICK=展示一键告警策略 NOT_ONECLICK=展示非一键告警策略
 	OneClickPolicyType []*string `json:"OneClickPolicyType,omitempty" name:"OneClickPolicyType"`
+
+	// 根据全部对象过滤，1代表需要过滤掉全部对象，0则无需过滤
+	NotBindAll *int64 `json:"NotBindAll,omitempty" name:"NotBindAll"`
+
+	// 根据实例对象过滤，1代表需要过滤掉有实例对象，0则无需过滤
+	NotInstanceGroup *int64 `json:"NotInstanceGroup,omitempty" name:"NotInstanceGroup"`
+
+	// 策略根据标签过滤
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 type DescribeAlarmPoliciesRequest struct {
@@ -3922,6 +3979,15 @@ type DescribeAlarmPoliciesRequest struct {
 
 	// 根据一键告警策略筛选 不传展示全部策略 ONECLICK=展示一键告警策略 NOT_ONECLICK=展示非一键告警策略
 	OneClickPolicyType []*string `json:"OneClickPolicyType,omitempty" name:"OneClickPolicyType"`
+
+	// 根据全部对象过滤，1代表需要过滤掉全部对象，0则无需过滤
+	NotBindAll *int64 `json:"NotBindAll,omitempty" name:"NotBindAll"`
+
+	// 根据实例对象过滤，1代表需要过滤掉有实例对象，0则无需过滤
+	NotInstanceGroup *int64 `json:"NotInstanceGroup,omitempty" name:"NotInstanceGroup"`
+
+	// 策略根据标签过滤
+	Tags []*Tag `json:"Tags,omitempty" name:"Tags"`
 }
 
 func (r *DescribeAlarmPoliciesRequest) ToJsonString() string {
@@ -3957,6 +4023,9 @@ func (r *DescribeAlarmPoliciesRequest) FromJsonString(s string) error {
 	delete(f, "NeedCorrespondence")
 	delete(f, "TriggerTasks")
 	delete(f, "OneClickPolicyType")
+	delete(f, "NotBindAll")
+	delete(f, "NotInstanceGroup")
+	delete(f, "Tags")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeAlarmPoliciesRequest has unknown keys!", "")
 	}
@@ -8262,6 +8331,31 @@ type IntegrationConfiguration struct {
 	GrafanaDashboardURL *string `json:"GrafanaDashboardURL,omitempty" name:"GrafanaDashboardURL"`
 }
 
+type LogAlarmReq struct {
+	// apm实例id
+	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
+
+	// 检索条件信息
+	Filter []*LogFilterInfo `json:"Filter,omitempty" name:"Filter"`
+
+	// 告警合并开启/暂停
+	AlarmMerge *string `json:"AlarmMerge,omitempty" name:"AlarmMerge"`
+
+	// 告警合并时间
+	AlarmMergeTime *string `json:"AlarmMergeTime,omitempty" name:"AlarmMergeTime"`
+}
+
+type LogFilterInfo struct {
+	// 字段名
+	Key *string `json:"Key,omitempty" name:"Key"`
+
+	// 比较符号
+	Operator *string `json:"Operator,omitempty" name:"Operator"`
+
+	// 字段值
+	Value *string `json:"Value,omitempty" name:"Value"`
+}
+
 type ManagementCommand struct {
 	// Agent 安装命令
 	// 注意：此字段可能返回 null，表示取不到有效值。
@@ -8536,6 +8630,9 @@ type ModifyAlarmPolicyConditionRequestParams struct {
 
 	// 聚合维度列表，指定按哪些维度 key 来做 group by
 	GroupBy []*string `json:"GroupBy,omitempty" name:"GroupBy"`
+
+	// 日志告警创建请求参数信息
+	LogAlarmReqInfo *LogAlarmReq `json:"LogAlarmReqInfo,omitempty" name:"LogAlarmReqInfo"`
 }
 
 type ModifyAlarmPolicyConditionRequest struct {
@@ -8561,6 +8658,9 @@ type ModifyAlarmPolicyConditionRequest struct {
 
 	// 聚合维度列表，指定按哪些维度 key 来做 group by
 	GroupBy []*string `json:"GroupBy,omitempty" name:"GroupBy"`
+
+	// 日志告警创建请求参数信息
+	LogAlarmReqInfo *LogAlarmReq `json:"LogAlarmReqInfo,omitempty" name:"LogAlarmReqInfo"`
 }
 
 func (r *ModifyAlarmPolicyConditionRequest) ToJsonString() string {
@@ -8582,6 +8682,7 @@ func (r *ModifyAlarmPolicyConditionRequest) FromJsonString(s string) error {
 	delete(f, "EventCondition")
 	delete(f, "Filter")
 	delete(f, "GroupBy")
+	delete(f, "LogAlarmReqInfo")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "ModifyAlarmPolicyConditionRequest has unknown keys!", "")
 	}
@@ -10041,6 +10142,14 @@ func (r *SetDefaultAlarmPolicyResponse) ToJsonString() string {
 // because it has no param check, nor strict type check
 func (r *SetDefaultAlarmPolicyResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
+}
+
+type Tag struct {
+	// 标签key
+	Key *string `json:"Key,omitempty" name:"Key"`
+
+	// 标签value
+	Value *string `json:"Value,omitempty" name:"Value"`
 }
 
 type TagInstance struct {