+++ testing/integration/import_test.go
@@ -221,6 +221,7 @@ import (
 	tdcpgv20211118 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/tdcpg/v20211118"
 	tdidv20210519 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/tdid/v20210519"
 	tdmqv20200217 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/tdmq/v20200217"
+	tdsv20220801 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/tds/v20220801"
 	temv20201221 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/tem/v20201221"
 	temv20210701 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/tem/v20210701"
 	teov20220106 "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/teo/v20220106"
@@ -2877,6 +2878,19 @@ func TestTdmqv20200217Import(t *testing.T) {
         profile.NewClientProfile())
     if err != nil || client == nil {
         t.Errorf("fail to init tdmq_v20200217 client: %v", err)
+    }
+}
+
+func TestTdsv20220801Import(t *testing.T) {
+    credential := common.NewCredential(
+        os.Getenv("TENCENTCLOUD_SECRET_ID"),
+        os.Getenv("TENCENTCLOUD_SECRET_KEY"))
+    client, err := tdsv20220801.NewClient(
+        credential,
+        regions.Guangzhou,
+        profile.NewClientProfile())
+    if err != nil || client == nil {
+        t.Errorf("fail to init tds_v20220801 client: %v", err)
     }
 }